+++ tests/Units/GroupEscalationTest.php
@@ -204,7 +204,7 @@ final class GroupEscalationTest extends EscaladeTestCase
                 'group_ticket' => 0,
             ],
             'update_expected' => [
-                'user_ticket' => 1,
+                'user_ticket' => 2,
                 'group_ticket' => 0,
             ],
         ];
@@ -258,7 +258,7 @@ final class GroupEscalationTest extends EscaladeTestCase
                 'group_ticket' => 1,
             ],
             'update_expected' => [
-                'user_ticket' => 1,
+                'user_ticket' => 2,
                 'group_ticket' => 1,
             ],
         ];
@@ -438,7 +438,7 @@ final class GroupEscalationTest extends EscaladeTestCase
                 'group_ticket' => 1,
             ],
             'update_expected' => [
-                'user_ticket' => 1,
+                'user_ticket' => 2,
                 'group_ticket' => 1,
             ],
         ];
@@ -703,8 +703,8 @@ final class GroupEscalationTest extends EscaladeTestCase
                 ],
             ]));
 
-            $this->assertEquals($provider['update_expected']['user_ticket'], count($ticket_user->find(['tickets_id' => $ticket_id, 'type' => CommonITILActor::ASSIGN])));
-            $this->assertEquals($provider['update_expected']['group_ticket'], count($group_ticket->find(['tickets_id' => $ticket_id, 'type' => CommonITILActor::ASSIGN])));
+            $this->assertEquals($provider['update_expected']['user_ticket'], count($ticket_user->find(['tickets_id' => $ticket->getID(), 'type' => CommonITILActor::ASSIGN])), 'Failed with config: ' . json_encode($provider['conf']));
+            $this->assertEquals($provider['update_expected']['group_ticket'], count($group_ticket->find(['tickets_id' => $ticket->getID(), 'type' => CommonITILActor::ASSIGN])), 'Failed with config: ' . json_encode($provider['conf']));
 
             if ($provider['conf']['use_assign_user_group_modification'] === 1) {
                 if ($provider['conf']['use_assign_user_group'] === 1) {